requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cycler==0.12.1
debugpy @ file:///D:/bld/debugpy_1744321277463/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dnspython==2.7.0
etelemetry==0.3.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.0
flask-cors==5.0.1
Flask-PyMongo==3.0.1
fonttools==4.57.0
frontend==0.0.3
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources==6.5.2
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1745672185/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
isodate==0.6.1
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1727163532151/work
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.3
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mpmath==1.3.0
multidict==6.4.3
narwhals==1.38.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.4
ollama==0.4.7
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pathlib==1.0.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
piexif==1.1.3
pillow==11.2.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1742485085/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
propcache==0.3.1
protobuf==6.30.2
prov==2.0.1
psutil @ file:///D:/bld/psutil_1740663127374/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
puremagic==1.29
pyarrow==20.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydot==4.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pymongo==4.12.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyWavelets==1.8.0
pywin32==307
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1743831235531/work
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
simplejson==3.20.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.46.2
streamlit==1.45.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.15.2
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tornado @ file:///D:/bld/tornado_1732615900055/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
traits==7.0.2
transformers==4.36.2
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
Werkzeug==3.1.3
yarl==1.20.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
